change multi_cls_v7pose and write Match_labels_images

+++ ProcessLib/images_labels_files_split.py
@@ -13,11 +13,11 @@ import shutil
 
 # Folder address where the original file is stored
 # total txt file path
-total_labels_path = "/media/hkuit164/Backup/yolov7/datasets/General_cls_0628/total_labels"
-target_floder_path = os.path.dirname(total_labels_path)
+total_labels_path = "/media/hkuit164/WD20EJRX/Aiden/Tennis_dataset/cls_sources/cls_sources_0710/relabel_json/relabel_cls_json"
+target_floder_path = "/media/hkuit164/WD20EJRX/Aiden/Tennis_dataset/cls_sources/cls_sources_0710/relabel_json/relabel_split_json"
 
 # total image file path
-total_imamges_path = "/media/hkuit164/WD20EJRX/Aiden/Tennis_dataset/cls_sources/cls_sources_0628/image_merge_cls_total/total_merge_images"
+total_imamges_path = "/media/hkuit164/WD20EJRX/Aiden/Tennis_dataset/cls_sources/cls_sources_0710/images/total_images"
 
 
 
@@ -60,14 +60,20 @@ def labels_split_proportion(total_labels_path, target_floder_path, train, val):
         train_folder_path = os.path.join(os.path.join(target_floder_path, "labels"), "train")
         if not os.path.exists(train_folder_path):
             os.makedirs(train_folder_path)
-        shutil.copy(old_path, train_folder_path)
+        if os.path.exists(old_path):
+            shutil.copy(old_path, train_folder_path)
+        else:
+            continue
 
     for fileName in val:
         old_path = os.path.join(total_labels_path, fileName)
         val_folder_path = os.path.join(os.path.join(target_floder_path, "labels"), "val")
         if not os.path.exists(val_folder_path):
             os.makedirs(val_folder_path)
-        shutil.copy(old_path, val_folder_path)
+        if os.path.exists(old_path):
+            shutil.copy(old_path, val_folder_path)
+        else:
+            continue
     return
 
 def images_split_proportion(total_imamges_path, target_floder_path, train, val, image_type):
@@ -77,29 +83,37 @@ def images_split_proportion(total_imamges_path, target_floder_path, train, val, 
         train_folder_path = os.path.join(os.path.join(target_floder_path, "images"), "train")
         if not os.path.exists(train_folder_path):
             os.makedirs(train_folder_path)
-        shutil.copy(old_path, train_folder_path)
+        if os.path.exists(old_path):
+            shutil.copy(old_path, train_folder_path)
+        else:
+            continue
 
     for fileName in val:
         old_path = os.path.join(total_imamges_path, fileName.split(".")[0] + "." + image_type)
         val_folder_path = os.path.join(os.path.join(target_floder_path, "images"), "val")
         if not os.path.exists(val_folder_path):
             os.makedirs(val_folder_path)
-        shutil.copy(old_path, val_folder_path)
+        if os.path.exists(old_path):
+            shutil.copy(old_path, val_folder_path)
+        else:
+            continue
     return
 
 def check_image_label():
     # Select the two folders you want to check.
-    imageName = os.listdir("/media/hkuit164/Backup/yolov7/datasets/General_cls_0628/images/train")
-    labelName = os.listdir("/media/hkuit164/Backup/yolov7/datasets/General_cls_0628/labels/train")
+    imageName = os.listdir("/media/hkuit164/WD20EJRX/Aiden/Tennis_dataset/pose_cls_sources/pose_cls_0718/labels/train")
+    labelName = os.listdir("/media/hkuit164/WD20EJRX/Aiden/Tennis_dataset/pose_cls_sources/pose_cls_0718/new_images/train")
     image_compare = []
     label_compare = []
     for image in imageName:
         image_compare.append(image.split(".")[0])
     for label in labelName:
         label_compare.append(label.split(".")[0])
-    print("images: ", image_compare)
-    print("labels: ", label_compare)
-    print("Different file: ", set(image_compare).difference(set(label_compare)))
+    print("images: ", set(image_compare))
+    print("length of images: ", len(image_compare))
+    print("labels: ", set(label_compare))
+    print("length of labels: ", len(label_compare))
+    print("Different file: ", list(set(image_compare).difference(set(label_compare))) + list(set(label_compare).difference(set(image_compare))))
     print("Number of different file: ", len(set(image_compare).difference(set(label_compare))))
 
 # choose split method
